feat(web): the web build linked a 1 GiB wall that one number served as both the reservation and the budget, and Chromium's six-pane scene died against it on a desktop that constructs 4 GiB -- the link flag is wasm32's 65,536 pages, every page and worker constructs the largest memory its engine accepts down a 4096..128 MiB ladder, and the budgets keep judging each heap against exactly the per-device policy they read before (1024/1024, or 512/256 by pointer type), so no budget, watermark, fit or admission figure moves on any device
The only behaviour change is that an allocation past the policy no longer
aborts while the reservation has room. An iPhone 13 Pro constructs 4 GiB and
iOS kills the tab near 2.3 GiB, so nothing is sized from the reservation
until the wall model measures one.

- heap.js: `LADDER_BYTES`, `constructMemory` (catches the constructor's throw
  and nothing else), `initWithHeap` walks down on a refusal or an engine
  LinkError and prints `linear memory ladder: constructed N MiB after refusing
  [..]` once; the `could not instantiate with a` warn is left only for an
  instance that took no rung. The pointer-type classification stays, renamed
  `choosePolicyHeapBytes`, as the policy figure alone.
- `start(page_reserved, page_policy, worker_policy)`; `heap_max` holds the two
  figures apart (`this_reservation` / `this_policy`, `worker_*`), a policy is
  held under its reservation, and the old accessor names are gone so an
  in-flight reader fails to compile rather than reading 4 GiB as a budget.
- `WASM_LINEAR_MEMORY_MAX_BYTES` splits into `WASM_LINKED_MAX_BYTES` (4 GiB)
  and `WASM_POLICY_HEAP_BYTES` (1 GiB, the wasm host presumption), for the
  same reason; `presumed_host_bytes` is u64.
- `budget state:` keeps `heap max a/b MiB` as the policy, with its meaning and
  values unchanged, and adds `reserved a/b MiB` inside the same comma group,
  so the separator count and every positional group are unmoved.
- calibrate.html reads this build's rungs and policy off heap.js; the beacon's
  `linear memory ladder:` needle is `written`.

+++ squallar-device-profile/src/budget/tests.rs
@@ -2610,7 +2610,7 @@ fn a_page_that_said_what_its_heap_was_built_with_outranks_the_bracket() {
     silent.limits = BudgetLimits::WASM;
     let linked = BudgetLimits::WASM
         .presumed_host_bytes
-        .expect("the wasm bracket declares a host presumption") as u64;
+        .expect("the wasm bracket declares a host presumption");
     assert_eq!(
         silent.capacity().host_bytes,
         Some(linked),
